incluir concluido e pendente

+++ Projeto/tarefas/src/main/java/com/example/tarefas/service/tarefa/TarefaService.java
@@ -3,11 +3,13 @@ package com.example.tarefas.service.tarefa;
 
 import com.example.tarefas.controller.tarefa.dto.TarefaDto;
 import com.example.tarefas.controller.usuario.dto.UsuarioDto;
+import com.example.tarefas.exceptions.BadRequestException;
 import com.example.tarefas.model.Tarefa;
 import com.example.tarefas.model.Usuario;
 import com.example.tarefas.repository.TarefaRepository;
 import com.example.tarefas.repository.UsuarioRepository;
 import jakarta.persistence.EntityNotFoundException;
+import jakarta.transaction.Transactional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Service;
@@ -19,11 +21,17 @@ import java.util.Optional;
 
 @Service
 public class TarefaService {
-    @Autowired
-    TarefaRepository tarefaRepository;
 
+    @Autowired
+    private TarefaRepository tarefaRepository;
+
+    @Autowired
+    private UsuarioRepository usuarioRepository;
+
+    @Transactional
     public List<Tarefa> findAll() { return tarefaRepository.findAll(); }
 
+    @Transactional
     public Tarefa findById(long id) {
         Optional<Tarefa> task  = tarefaRepository.findById(id);
         if(task.isPresent()) {
@@ -36,8 +44,7 @@ public class TarefaService {
 
 
 
-    @Autowired
-    private UsuarioRepository usuarioRepository;
+
 
     public Tarefa create(TarefaDto tarefaDto) {
         Usuario usuario = usuarioRepository.findById(tarefaDto.usuarioId())
@@ -53,8 +60,13 @@ public class TarefaService {
         return tarefaRepository.save(tarefa);
     }
 
-    public Tarefa update(Long id, TarefaDto tarefaDto) {
+    public Tarefa update(Long id, TarefaDto tarefaDto, Long usuarioId) {
         Tarefa tarefaExistente = findById(id);
+
+        if (!tarefaExistente.getUsuario_criado().getId().equals(tarefaDto.usuarioId())) {
+            throw new BadRequestException("Você não tem permissão para alterar esta tarefa.");
+        }
+
 
         Usuario usuario = usuarioRepository.findById(tarefaDto.usuarioId())
                 .orElseThrow(() -> new RuntimeException("Usuário não encontrado"));
@@ -69,15 +81,35 @@ public class TarefaService {
         return tarefaRepository.save(tarefaAtualizada);
     }
 
-    public Tarefa alternarConclusao(Long id) {
+    public Tarefa concluirTarefa(Long id, Long usuarioId) {
         Tarefa tarefa = findById(id);
-        tarefa.setConcluido(!Boolean.TRUE.equals(tarefa.getConcluido()));
+
+        if (!tarefa.getUsuario_criado().getId().equals(usuarioId)) {
+            throw new BadRequestException("Você não tem permissão para concluir esta tarefa.");
+        }
+
+        tarefa.setConcluido(true);
+        return tarefaRepository.save(tarefa);
+    }
+
+    public Tarefa pendenteTarefa(Long id, Long usuarioId) {
+        Tarefa tarefa = findById(id);
+
+        if (!tarefa.getUsuario_criado().getId().equals(usuarioId)) {
+            throw new BadRequestException("Você não tem permissão para alterar esta tarefa.");
+        }
+
+        tarefa.setConcluido(false);
         return tarefaRepository.save(tarefa);
     }
 
 
-    public void delete(Long id) {
-        findById(id);
+
+    public void delete(Long id, Long usuarioId) {
+        Tarefa tarefaExistente = findById(id);
+        if (!tarefaExistente.getUsuario_criado().getId().equals(usuarioId)) {
+            throw new BadRequestException("Você não tem permissão para deletar esta tarefa.");
+        }
         tarefaRepository.deleteById(id);
     }
